+++ src/test/converter2/issues/gh2779.ts
@@ -1,0 +1,8 @@
+declare module "bar" {
+    export namespace Nested {
+        export interface Bar {}
+    }
+}
+declare module "foo" {
+    export type Foo = import("bar").Nested.Bar;
+}